// libraries/api-client/src/openapi.d.ts
export interface paths {
    '/v1/liveness': {
        get: operations['v1.liveness']
    }
    '/v1/keys.getKey': {
        get: operations['getKey']
    }
    '/v1/keys.whoami': {
        post: operations['whoami']
    }
    '/v1/keys.deleteKey': {
        post: operations['deleteKey']
    }
    '/v1/keys.createKey': {
        post: operations['createKey']
    }
    '/v1/keys.verifyKey': {
        post: operations['verifyKey']
    }
    '/v1/keys.updateKey': {
        post: operations['updateKey']
    }
    '/v1/keys.updateRemaining': {
        post: operations['updateRemaining']
    }
    '/v1/keys.getVerifications': {
        get: operations['keys.getVerifications']
    }
    '/v1/keys.addPermissions': {
        post: operations['addPermissions']
    }
    '/v1/keys.removePermissions': {
        post: operations['removePermissions']
    }
    '/v1/keys.setPermissions': {
        post: operations['setPermissions']
    }
    '/v1/keys.addRoles': {
        post: operations['addRoles']
    }
    '/v1/keys.removeRoles': {
        post: operations['removeRoles']
    }
    '/v1/keys.setRoles': {
        post: operations['setRoles']
    }
    '/v1/apis.getApi': {
        get: operations['getApi']
    }
    '/v1/apis.createApi': {
        post: operations['createApi']
    }
    '/v1/apis.listKeys': {
        get: operations['listKeys']
    }
    '/v1/apis.deleteApi': {
        post: operations['deleteApi']
    }
    '/v1/apis.deleteKeys': {
        post: operations['deleteKeys']
    }
    '/v1/ratelimits.limit': {
        post: operations['limit']
    }
    '/v1/ratelimits.setOverride': {
        post: operations['setOverride']
    }
    '/v1/ratelimits.listOverrides': {
        get: operations['listOverrides']
    }
    '/v1/ratelimits.deleteOverride': {
        post: operations['deleteOverride']
    }
    '/v1/ratelimits.getOverride': {
        get: operations['getOverride']
    }
    '/v1/migrations.createKeys': {
        post: operations['v1.migrations.createKeys']
    }
    '/v1/migrations.enqueueKeys': {
        post: operations['v1.migrations.enqueueKeys']
    }
    '/v1/permissions.createPermission': {
        post: operations['createPermission']
    }
    '/v1/permissions.deletePermission': {
        post: operations['deletePermission']
    }
    '/v1/permissions.getPermission': {
        get: operations['getPermission']
    }
    '/v1/permissions.listPermissions': {
        get: operations['listPermissions']
    }
    '/v1/permissions.createRole': {
        post: operations['createRole']
    }
    '/v1/permissions.deleteRole': {
        post: operations['deleteRole']
    }
    '/v1/permissions.getRole': {
        get: operations['getRole']
    }
    '/v1/permissions.listRoles': {
        get: operations['listRoles']
    }
    '/v1/identities.createIdentity': {
        post: operations['createIdentity']
    }
    '/v1/identities.getIdentity': {
        get: operations['getIdentity']
    }
    '/v1/identities.listIdentities': {
        get: operations['listIdentities']
    }
    '/v1/identities.updateIdentity': {
        post: operations['updateIdentity']
    }
    '/v1/identities.deleteIdentity': {
        post: operations['deleteIdentity']
    }
    '/v1/analytics.getVerifications': {
        get: operations['getVerifications']
    }
    '/v1/keys': {
        post: operations['deprecated.createKey']
    }
    '/v1/keys/verify': {
        /** @deprecated */
        post: operations['deprecated.verifyKey']
    }
    '/v1/apis/{apiId}/keys': {
        get: operations['deprecated.listKeys']
    }
}

export type webhooks = Record<string, never>

export interface components {
    schemas: {
        ErrBadRequest: {
            error: {
                /**
                 * @description A machine readable error code.
                 * @example BAD_REQUEST
                 * @enum {string}
                 */
                code: 'BAD_REQUEST'
                /**
                 * @description A link to our documentation with more details about this error code
                 * @example https://unkey.dev/docs/api-reference/errors/code/BAD_REQUEST
                 */
                docs: string
                /** @description A human readable explanation of what went wrong */
                message: string
                /**
                 * @description Please always include the requestId in your error report
                 * @example req_1234
                 */
                requestId: string
            }
        }
// ...
    }
    responses: never
    parameters: {}
    requestBodies: never
    headers: never
    pathItems: never
}

export type $defs = Record<string, never>

export type external = Record<string, never>

export interface operations {
    'v1.liveness': {
        responses: {
            /** @description The configured services and their status */
            200: {
                content: {
                    'application/json': {
                        /** @description The status of the server */
                        status: string
                        services: {
                            /**
                             * @description The name of the connected metrics service
                             * @example AxiomMetrics
                             */
                            metrics: string
                            /**
                             * @description The name of the connected logger service
                             * @example AxiomLogger or ConsoleLogger
                             */
                            logger: string
                            /** @description The name of the connected ratelimit service */
                            ratelimit: string
                            /** @description The name of the connected usagelimit service */
                            usagelimit: string
                        }
                    }
                }
            }
            /** @description The server cannot or will not process the request due to something that is perceived to be a client error (e.g., malformed request syntax, invalid request message framing, or deceptive request routing). */
            400: {
                content: {
                    'application/json': components['schemas']['ErrBadRequest']
                }
            }
            /** @description Although the HTTP standard specifies "unauthorized", semantically this response means "unauthenticated". That is, the client must authenticate itself to get the requested response. */
            401: {
                content: {
                    'application/json': components['schemas']['ErrUnauthorized']
                }
            }
            /** @description The client does not have access rights to the content; that is, it is unauthorized, so the server is refusing to give the requested resource. Unlike 401 Unauthorized, the client's identity is known to the server. */
            403: {
                content: {
                    'application/json': components['schemas']['ErrForbidden']
                }
            }
            /** @description The server cannot find the requested resource. In the browser, this means the URL is not recognized. In an API, this can also mean that the endpoint is valid but the resource itself does not exist. Servers may also send this response instead of 403 Forbidden to hide the existence of a resource from an unauthorized client. This response code is probably the most well known due to its frequent occurrence on the web. */
            404: {
                content: {
                    'application/json': components['schemas']['ErrNotFound']
                }
            }
            /** @description This response is sent when a request conflicts with the current state of the server. */
            409: {
                content: {
                    'application/json': components['schemas']['ErrConflict']
                }
            }
            /** @description The requested operation cannot be completed because certain conditions were not met. This typically occurs when a required resource state or version check fails. */
            412: {
                content: {
                    'application/json': components['schemas']['ErrPreconditionFailed']
                }
            }
            /** @description The user has sent too many requests in a given amount of time ("rate limiting") */
            429: {
                content: {
                    'application/json': components['schemas']['ErrTooManyRequests']
                }
            }
            /** @description The server has encountered a situation it does not know how to handle. */
            500: {
                content: {
                    'application/json': components['schemas']['ErrInternalServerError']
                }
            }
        }
    }
    getKey: {
        parameters: {
            query: {
                keyId: string
                decrypt?: boolean | null
            }
        }
// ...
    }
    whoami: {
        requestBody: {
            content: {
                'application/json': {
                    /**
                     * @description The actual key to fetch
                     * @example sk_123
                     */
                    key: string
                }
            }
        }
// ...
    }
// ...
    'v1.migrations.enqueueKeys': {
// ...
        responses: {
            /** @description The key ids of all created keys */
            202: {
                content: {
                    'application/json': Record<string, never>
                }
            }
// ...
        }
    }
// ...
    'deprecated.verifyKey': {
// ...
        responses: {
            /** @description The verification result */
            200: {
                content: {
                    'application/json': {
// ...
                        code?:
                            | 'NOT_FOUND'
                            | 'FORBIDDEN'
                            | 'USAGE_EXCEEDED'
                            | 'RATE_LIMITED'
                            | 'UNAUTHORIZED'
                            | 'DISABLED'
                            | 'INSUFFICIENT_PERMISSIONS'
                            | 'EXPIRED'
                    }
                }
            }
// ...
        }
    }
// ...
}